Refactor for high cohesion, low coupling, and testability
Split ontology.py (1162 lines) into focused sub-package:
- ontology/crud.py: CRUD operations (list, get, create, delete)
- ontology/builder.py: OntologyBuilder, Property, EntityType, etc.
- ontology/parts.py: low-level definition parts manipulation
- ontology/_id_gen.py: ID generation utilities
- ontology/__init__.py: re-exports all public symbols

Fix GraphClient private method access:
- Add FabricClient.raw_request() public method for custom LRO patterns
- GraphClient.refresh() now uses raw_request() instead of _request()

Make FabricClient configurable for testability:
- Constructor accepts base_url and timeout keyword parameters
- Tests can point at mock API servers without patching

Remove global mutable state from auth.py:
- Delete _default_credential, _get_default(), module-level get_token()
- ensure_logged_in() creates fresh credentials (no global state)
- Eliminates test isolation issues

Consolidate duplicate encode_part:
- bundle.py imports encode_part from crud.py directly
- Single source of truth for API part encoding

290 tests pass, all lint/mypy/format clean.
