+++ pyproject.toml
@@ -136,6 +136,7 @@ module = [
     "pyfabric.client.graph",
     "pyfabric.client.livy",
     "pyfabric.client.ontology",
+    "pyfabric.client.ontology.*",
     "pyfabric.client.ontology_sync",
     "pyfabric.data.*",
     "pyfabric.workspace.*",
